debug

+++ 2017/dnn_rl.py
@@ -140,7 +140,7 @@ def q_pretrain(x_path, noise_path, cluster_path, model_path,
         l1.apply(weights)
     
         l1.cuda()
-        l1 = l1.to(device)
+        l1 = l1.to(device).double()
         criterion.cuda()
 
         optimizer = optim.SGD(l1.parameters(), lr=0.001, momentum=0.9)
